+++ changelog.md
@@ -275,6 +275,8 @@
 - Added `genasts.genAst` that avoids the problems inherent with `quote do` and can
   be used as a replacement.
 
+- Added `std/syntaxes` to hint syntax highlighters about string literals.
+
 ## Language changes
 
 - `nimscript` now handles `except Exception as e`.